Make test run with babel-node directly without seprate build-test command
