+++ .travis.yml
@@ -1,6 +1,7 @@
 language: node_js
 node_js:
   - "stable"
+sudo: false # makes tests faster
 install:
   - npm install
 script: npm run pretest
