+++ package.json
@@ -4,12 +4,11 @@
   "description": "Server side app that provides mydiary ui client with data",
   "main": "index.js",
   "scripts": {
-    "test": "mocha",
+    "test": "babel-node node_modules/mocha/bin/_mocha test/index.js",
     "pretest": "eslint .",
     "lintfix": "eslint . --fix",
-    "build": "babel src -d dist",
-    "build-test": "babel src-test -d test",
-    "cover": "istanbul cover --hook-run-in-context node_modules/mocha/bin/_mocha -- -R spec --timeout 10000 test/index.js",
+    "build": "npm run lintfix && babel src -d dist",
+    "cover": "babel-node ./node_modules/istanbul/lib/cli cover --hook-run-in-context node_modules/mocha/bin/_mocha -- -R spec --timeout 10000 test/index.js",
     "coveralls": "npm run cover -- --report lcovonly && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
     "dev": "nodemon ./dist/server",
     "start": "node ./dist/server"
